fix: remove duplicate session_event_handler definitions
