+++ src/adapters/inbound/http/mod.rs
@@ -1,3 +1,2 @@
 pub mod dto;
-pub mod routes;
 pub mod time_metrics_adapter;
